plla_tisvs/requirements.txt
Package            Version
------------------ ------------
appdirs            1.4.4
attrs              21.2.0
audioread          2.1.9
certifi            2021.5.30
cffi               1.14.6
charset-normalizer 2.0.6
cycler             0.10.0
decorator          5.1.0
ffmpeg             1.4
ffmpeg-python      0.2.0
future             0.18.2
idna               3.2
importlib-metadata 4.8.1
joblib             1.0.1
jsonschema         3.2.0
kiwisolver         1.3.2
librosa            0.8.1
llvmlite           0.37.0
matplotlib         3.4.3
musdb              0.4.0
museval            0.4.0
norbert            0.2.1
numba              0.54.0
numpy              1.20.3
packaging          21.0
pandas             1.3.3
Pillow             8.3.2
pip                21.0.1
pooch              1.5.1
praat-parselmouth  0.4.0
protobuf           3.18.0
pyaml              21.8.3
pycparser          2.20
pyparsing          2.4.7
pyrsistent         0.18.0
python-dateutil    2.8.2
pytz               2021.1
PyYAML             5.4.1
requests           2.26.0
resampy            0.2.2
scikit-learn       1.0
scipy              1.7.1
seaborn            0.11.2
setuptools         58.0.4
simplejson         3.17.5
six                1.16.0
SoundFile          0.10.3.post1
stempeg            0.2.3
tensorboardX       2.4
threadpoolctl      2.2.0
torch              1.9.1
tqdm               4.62.3
typing-extensions  3.10.0.2
urllib3            1.26.7
wheel              0.37.0
zipp               3.5.1
